set up basic auth

+++ src/posts/views.py
@@ -1,5 +1,5 @@
 from django.contrib import messages
-from django.http import HttpResponse, HttpResponseRedirect
+from django.http import HttpResponse, HttpResponseRedirect, Http404
 from django.shortcuts import render, get_object_or_404, redirect
 
 # Create your views here.
@@ -7,9 +7,14 @@ from .forms import PostForm
 from .models import Post
 
 def post_create(request):
+    if not request.user.is_staff:
+        raise Http404
+    if not request.user.is_authenticated():
+        raise Http404
     form = PostForm(request.POST or None, request.FILES or None)
     if form.is_valid():
         instance = form.save(commit=False)
+        instance.user = request.user
         instance.save()
         messages.success(request, "Successfuly Created")
         return HttpResponseRedirect(instance.get_absolute_url())
@@ -35,6 +40,8 @@ def post_list(request):
     return render(request, "post_list.html", context)
 
 def post_update(request, slug=None):
+    if not request.user.is_staff:
+        raise Http404
     instance = get_object_or_404(Post, slug=slug)
     form = PostForm(request.POST or None, request.FILES or None, instance=instance)
     if form.is_valid():
@@ -51,6 +58,8 @@ def post_update(request, slug=None):
 
 
 def post_delete(request, slug=None):
+    if not request.user.is_staff:
+        raise Http404
     instance = get_object_or_404(Post, slug=slug)
     instance.delete()
     messages.success(request, "Successfuly Deleted")
